requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 24.11.3
absl-py=2.1.0=pypi_0
astunparse=1.6.3=pypi_0
backcall=0.2.0=pypi_0
ca-certificates=2024.12.31=haa95532_0
cachetools=5.5.0=pypi_0
certifi=2024.8.30=pyhd8ed1ab_0
charset-normalizer=3.4.1=pypi_0
colorama=0.4.6=pypi_0
cudatoolkit=11.2.2=h7d7167e_13
cudnn=8.1.0.77=h3e0f4f4_0
cycler=0.11.0=pypi_0
decorator=5.1.1=pypi_0
flatbuffers=24.12.23=pypi_0
fonttools=4.38.0=pypi_0
gast=0.4.0=pypi_0
google-auth=2.37.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.62.3=pypi_0
h5py=3.8.0=pypi_0
idna=3.10=pypi_0
imageio=2.31.2=pypi_0
imgaug=0.4.0=pypi_0
importlib-metadata=6.7.0=pypi_0
ipython=7.34.0=pypi_0
jedi=0.19.2=pypi_0
keras=2.10.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.4.5=pypi_0
lapx=0.5.11.post1=pypi_0
libclang=18.1.1=pypi_0
markdown=3.4.4=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.5.3=pypi_0
matplotlib-inline=0.1.6=pypi_0
networkx=2.6.3=pypi_0
numpy=1.21.6=pypi_0
oauthlib=3.2.2=pypi_0
opencv-python=4.10.0.84=pypi_0
openssl=3.4.0=ha4e3fda_1
opt-einsum=3.3.0=pypi_0
packaging=24.0=pypi_0
pandas=1.3.5=pypi_0
parso=0.8.4=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.5.0=pypi_0
pip=22.3.1=py37haa95532_0
prompt-toolkit=3.0.48=pypi_0
protobuf=3.19.6=pypi_0
psutil=6.1.1=pypi_0
py-cpuinfo=9.0.0=pypi_0
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pygments=2.17.2=pypi_0
pyparsing=3.1.4=pypi_0
python=3.7.12=h900ac77_100_cpython
python-dateutil=2.9.0.post0=pypi_0
pytz=2024.2=pypi_0
pywavelets=1.3.0=pypi_0
pyyaml=6.0.1=pypi_0
requests=2.31.0=pypi_0
requests-oauthlib=2.0.0=pypi_0
rsa=4.9=pypi_0
scikit-image=0.19.3=pypi_0
scipy=1.7.3=pypi_0
seaborn=0.12.2=pypi_0
setuptools=65.6.3=py37haa95532_0
shapely=2.0.6=pypi_0
six=1.17.0=pypi_0
sqlite=3.45.3=h2bbff1b_0
tensorboard=2.10.1=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow-estimator=2.10.0=pypi_0
tensorflow-gpu=2.10.0=pypi_0
tensorflow-io-gcs-filesystem=0.31.0=pypi_0
termcolor=2.3.0=pypi_0
tifffile=2021.11.2=pypi_0
torch=1.13.1=pypi_0
torchvision=0.14.1=pypi_0
tqdm=4.67.1=pypi_0
traitlets=5.9.0=pypi_0
typing-extensions=4.7.1=pypi_0
ucrt=10.0.22621.0=h57928b3_1
ultralytics=8.0.145=pypi_0
urllib3=2.0.7=pypi_0
vc=14.40=haa95532_2
vc14_runtime=14.42.34433=he29a5d6_23
vs2015_runtime=14.42.34433=hdffcdeb_23
wcwidth=0.2.13=pypi_0
werkzeug=2.2.3=pypi_0
wheel=0.38.4=py37haa95532_0
wincertstore=0.2=py37haa95532_2
wrapt=1.16.0=pypi_0
zipp=3.15.0=pypi_0
